% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_parallel.R
\name{cat_parallel}
\alias{cat_parallel}
\alias{message_parallel}
\title{Versions of \code{\link[=cat]{cat()}} and \code{\link[=message]{message()}} for parallel processing}
\usage{
cat_parallel(...)

message_parallel(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character and which
are pasted together.}
}
\value{
Prints a message to the console. \code{cat_parallel()} uses no line feed,
while \code{message_parallel()} always adds a newline.
}
\description{
Prints messages to the console using \code{echo} during to enable messages to be
printed during parallel processing. Text is only printed if the Rstudio
environment is detected.
}
