% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_parallel.R
\name{catchError}
\alias{catchError}
\title{Catch error messages during parallel processing}
\usage{
catchError(expr, ...)
}
\arguments{
\item{expr}{An expression to be wrapped in \code{\link[=try]{try()}} to allow execution and
catch error messages.}

\item{...}{Optional objects to be tracked if you want to know state of
objects at the point error messages are generated.}
}
\value{
Prints error messages during parallel processing. If there is no
error, the result of the evaluated expression is returned.
}
\description{
Allows an expression to be wrapped in \code{\link[=try]{try()}} to catch error messages. Any
error messages are printed to the console using \code{\link[=mcstop]{mcstop()}}.
}
\seealso{
\code{\link[=mcstop]{mcstop()}}
}
