% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_parallel.R
\name{mcstop}
\alias{mcstop}
\title{Stop and print error message during parallel processing}
\usage{
mcstop(...)
}
\arguments{
\item{...}{Objects coerced to character and pasted together and printed to
the console using \code{echo}.}
}
\value{
Prints an error message.
}
\description{
\code{mcstop()} is a multicore version of \code{\link[=stop]{stop()}} which prints to the console
using 'echo' during parallel commands such as \code{\link[=mclapply]{mclapply()}}, to allow error
messages to be more visible.
}
